+++ src/e_topdown.f95
@@ -28,44 +28,6 @@ RETURN
 END ! # nocov end
 
 
-!-------------------------------------------------------------------
-! parallel starts
-
-SUBROUTINE emistd1fpar (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(nrowv, ncolv)
-DOUBLE PRECISION :: month(pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt, ntmax
-
-ntmax =  OMP_GET_MAX_THREADS()
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-        emis(i, j, k) = veh(i, j) * lkm(j) * ef(i, j)*month(k)
-      ENDDO
-   ENDDO
-ENDDO
-
-CALL OMP_SET_NUM_THREADS(ntmax)
-RETURN
-END ! # nocov end
-
 SUBROUTINE emistd2f (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis) ! # nocov startz
 IMPLICIT NONE
 
@@ -103,44 +65,6 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
-
-SUBROUTINE emistd2fpar (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv,ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(nrowv,ncolv)
-DOUBLE PRECISION :: month(nrowv, pmonth)
-DOUBLE PRECISION :: emis(nrowv,ncolv,pmonth)
-
-INTEGER i, j, k, nt, ntmax
-
-ntmax =  OMP_GET_MAX_THREADS()
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-        emis(i, j, k) = veh(i, j) * lkm(j) * ef(i, j)*month(i, k)
-      ENDDO
-   ENDDO
-ENDDO
-
-CALL OMP_SET_NUM_THREADS(ntmax)
-
-RETURN
-END ! # nocov end
 
 !-------------------------------------------------------------------
 ! cold
@@ -169,38 +93,6 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
-
-SUBROUTINE emistd2coldfpar (nrowv, ncolv, veh, lkm, ef, efcold, beta, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(ncolv)
-DOUBLE PRECISION :: efcold(nrowv, ncolv)
-DOUBLE PRECISION :: beta(nrowv)
-DOUBLE PRECISION :: emis(nrowv, ncolv)
-
-INTEGER i, j, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      emis(i, j) = beta(i) * veh(i, j) * lkm(j) * ef(j) * efcold(i, j)
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 SUBROUTINE emistd3f (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis) ! # nocov start
 
@@ -228,40 +120,6 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
-
-SUBROUTINE emistd3fpar (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(nrowv, ncolv, pmonth)
-DOUBLE PRECISION :: month(pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-        emis(i, j, k) = veh(i, j) * lkm(j) * ef(i, j, k) * month(k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 !-------------------------------------------------------------------
 ! cold
@@ -294,42 +152,7 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd3coldfpar (nrowv, ncolv, pmonth, veh, lkm, ef, efcold, beta, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(ncolv)
-DOUBLE PRECISION :: efcold(nrowv, ncolv, pmonth)
-DOUBLE PRECISION :: beta(nrowv, pmonth)
-DOUBLE PRECISION :: month(pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-         emis(i, j, k) = beta(i, k) * veh(i, j) * lkm(j) * ef(j) * efcold(i, j, k) * month(k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 SUBROUTINE emistd4f (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis) ! # nocov start
 
@@ -357,40 +180,7 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd4fpar (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(nrowv, ncolv, pmonth)
-DOUBLE PRECISION :: month(nrowv, pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-         emis(i, j, k) = veh(i, j) * lkm(j) * ef(i, j, k) * month(i, k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 !-------------------------------------------------------------------
 ! cold
@@ -423,42 +213,7 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd4coldfpar (nrowv, ncolv, pmonth, veh, lkm, ef, efcold, beta, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(nrowv, ncolv)
-DOUBLE PRECISION :: efcold(nrowv, ncolv, pmonth)
-DOUBLE PRECISION :: beta(nrowv, pmonth)
-DOUBLE PRECISION :: month(nrowv, pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-         emis(i, j, k) = beta(i, k) * veh(i, j) * lkm(j) * ef(i, j) * efcold(i, j, k) * month(i, k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 SUBROUTINE emistd5f (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis) ! # nocov start
 
@@ -486,40 +241,7 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd5fpar (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(ncolv)
-DOUBLE PRECISION :: month(pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-        emis(i, j, k) = veh(i, j) * lkm(j) * ef(j) * month(k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 !-------------------------------------------------------------------
 ! cold
@@ -552,42 +274,7 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd5coldfpar (nrowv, ncolv, pmonth, veh, lkm, ef, efcold, beta, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(nrowv, ncolv)
-DOUBLE PRECISION :: efcold(nrowv, ncolv, pmonth)
-DOUBLE PRECISION :: beta(nrowv, pmonth)
-DOUBLE PRECISION :: month(pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-         emis(i, j, k) = beta(i, k) * veh(i, j) * lkm(j) * ef(i, j) * efcold(i, j, k) * month(k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 
 SUBROUTINE emistd6f (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis) ! # nocov start
@@ -616,40 +303,6 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
-
-SUBROUTINE emistd6fpar (nrowv, ncolv, pmonth, veh, lkm, ef, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(ncolv)
-DOUBLE PRECISION :: month(nrowv,pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-        emis(i, j, k) = veh(i,j) * lkm(j) * ef(j) * month(i, k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 !-------------------------------------------------------------------
 ! cold
@@ -682,42 +335,7 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd6coldfpar (nrowv, ncolv, pmonth, veh, lkm, ef, efcold, beta, month, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-INTEGER pmonth
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(ncolv)
-DOUBLE PRECISION :: efcold(nrowv, ncolv, pmonth)
-DOUBLE PRECISION :: beta(nrowv, pmonth)
-DOUBLE PRECISION :: month(nrowv, pmonth)
-DOUBLE PRECISION :: emis(nrowv, ncolv, pmonth)
-
-INTEGER i, j, k, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j, k) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-      DO k = 1, pmonth
-         emis(i, j, k) = beta(i, k) * veh(i, j) * lkm(j) * ef(j) * efcold(i, j, k) * month(i, k)
-      ENDDO
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end
 
 SUBROUTINE emistd7f (nrowv, ncolv, veh, lkm, ef, emis) ! # nocov start
 
@@ -741,33 +359,4 @@ ENDDO
 RETURN
 END ! # nocov end
 
-!-------------------------------------------------------------------
-! parallel starts
 
-SUBROUTINE emistd7fpar (nrowv, ncolv, veh, lkm, ef, emis, nt) ! # nocov start
-USE OMP_LIB
-IMPLICIT NONE
-
-INTEGER nrowv
-INTEGER ncolv
-DOUBLE PRECISION :: veh(nrowv, ncolv)
-DOUBLE PRECISION :: lkm(ncolv)
-DOUBLE PRECISION :: ef(ncolv)
-DOUBLE PRECISION :: emis(nrowv, ncolv)
-
-INTEGER i, j, nt
-
-
-CALL OMP_SET_DYNAMIC(.TRUE.)
-
-emis = 0.0
-
-!$OMP PARALLEL DO PRIVATE(i, j) DEFAULT(shared) NUM_THREADS(nt)
-DO i = 1, nrowv
-   DO j = 1, ncolv
-       emis(i, j) = veh(i,j) * lkm(j) * ef(j)
-   ENDDO
-ENDDO
-
-RETURN
-END ! # nocov end